+++ packages/vue-next/example/config.js
@@ -1,49 +1,6 @@
----
-layout: doc
-outline: deep
----
+import { request } from './request.js'
 
-# Quick Start
-
-## 安装
-
-```bash
-npm i @tinyuploader/vuenext -S
-```
-
-## 使用
-
-```javascript
-// main.js
-import { createApp } from 'vue'
-import App from './App.vue'
-import Uploader from '@tinyuploader/vuenext'
-import '@tinyuploader/vuenext/dist/style.css'
-
-const app = createApp(App)
-
-app.use(Uploader)
-app.mount('#app')
-```
-
-```vue
-<template>
-  <Uploader
-    v-bind="customOption"
-    :defaultFileList="defaultFileList"
-    ref="uploaderRef"
-    @onChange="onChange"
-  />
-</template>
-
-<script setup>
-import { ref, onMounted } from 'vue'
-import { customOption } from './config'
-
-const uploaderRef = ref()
-const defaultFileList = ref([])
-const fileList = ref([])
-const customOption = {
+export const customOption = {
   // input 属性相关
   drag: true,
   accept: '*',
@@ -60,7 +17,7 @@ const customOption = {
     }
   },
   beforeRemove: (_file) => true,
-  addFailToRemove: false, // 默认为true
+  addFailToRemove: false, // 默认为false
   chunkSize: 2 * 1024 * 1024, // 2M
   fakeProgres: true,
   withHash: true,
@@ -108,20 +65,3 @@ const customOption = {
   processData: (data, _processType) => data,
   customStatus: {}
 }
-
-onMounted(() => {
-  setTimeout(() => {
-    defaultFileList.value = [{ name: 'default.png', url: 'https://baidu.com', id: '1' }]
-  }, 300)
-})
-const onChange = (_file, list) => {
-  fileList.value = list
-}
-</script>
-```
-
-属性、方法用法参考[`@tinyuploader/vue`](/vue/quick-start)
-
-## [demo](https://codesandbox.io/p/devbox/dtv5l3)
-
-## [更新日志](https://github.com/moyuderen/uploader/blob/main/packages/vue-next/CHANGELOG.md)