+++ packages/vue-next/src/components/uploader-button.vue
@@ -1,0 +1,13 @@
+<template>
+  <span class="tiny-uploader-button">
+    <slot></slot>
+  </span>
+</template>
+
+<script setup></script>
+
+<style lang="scss">
+.tiny-uploader-button {
+  color: var(--tiny-color-primary);
+}
+</style>