+++ packages/vue-next/src/icons/remove-icon.vue
@@ -1,0 +1,29 @@
+<template>
+  <svg
+    t="1747993419307"
+    class="icon"
+    viewBox="0 0 1024 1024"
+    version="1.1"
+    xmlns="http://www.w3.org/2000/svg"
+    p-id="26718"
+    :width="size"
+    :height="size"
+  >
+    <path
+      d="M644.608 307.2L512 439.808 379.392 307.2 307.2 379.392 439.808 512 307.2 644.608 379.392 716.8 512 584.192 644.608 716.8 716.8 644.608 584.192 512 716.8 379.392 644.608 307.2zM512 0C228.864 0 0 228.864 0 512s228.864 512 512 512 512-228.864 512-512S795.136 0 512 0z m0 921.6c-225.792 0-409.6-183.808-409.6-409.6s183.808-409.6 409.6-409.6 409.6 183.808 409.6 409.6-183.808 409.6-409.6 409.6z"
+      fill="#f56c6c"
+      p-id="26719"
+    ></path>
+  </svg>
+</template>
+
+<script setup>
+import { toRefs } from 'vue'
+const props = defineProps({
+  size: {
+    type: Number,
+    default: 14
+  }
+})
+const { size } = toRefs(props)
+</script>